refactor: move fuzzy matching module
